[AOOS-679] Adds sophisticated processing to yield sea names for trajectory
This commit adds a new function to the geo module for producing the list
of sea names that intersect a trajectory. The computational approach
used to make the determination makes a somewhat balanced trade-off of
memory consumption for shortening the processing time. The new function
includes a specific unit test along with serialized test data from a
deployment for ATN.